dags/requirements.txt
-i https://pypi.org/simple
aiobotocore==1.3.3
aiohttp==3.7.4.post0
aioitertools==0.8.0
airflow-sqlcmd-operator==0.3.0
alembic==1.6.5
altgraph==0.17
anyio==3.3.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-microsoft-mssql==1.1.0
apache-airflow-providers-mysql==2.1.1
apache-airflow-providers-sqlite==2.0.0
apache-airflow==2.1.2
apispec[yaml]==3.3.2
argcomplete==1.12.3
async-timeout==3.0.1
attrs==20.3.0
authlib==0.15.4
babel==2.9.1
bcrypt==3.2.0
blinker==1.4
boto3==1.17.106
boto==2.49.0
botocore==1.20.106
cached-property==1.5.2
cattrs==1.5.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.0.2
commonmark==0.9.1
configparser==3.5.3
connexion==2.9.0
croniter==0.3.37
cryptography==3.4.8
datetime==4.3
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
et-xmlfile==1.1.0
flask-admin==1.5.4
flask-appbuilder==3.3.2
flask-babel==1.0.0
flask-caching==1.10.1
flask-jwt-extended==3.25.1
flask-login==0.4.1
flask-openid==1.2.5
flask-sqlalchemy==2.5.1
flask-swagger==0.2.14
flask-wtf==0.14.3
flask==1.1.4
fsspec==2021.7.0
funcsigs==1.0.2
future==0.18.2
graphviz==0.17
gunicorn==19.10.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
importlib-metadata==4.8.1 ; python_version < '3.9'
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
macholib==1.15
mako==1.1.4
markdown==2.6.11
markupsafe==1.1.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
marshmallow==3.13.0
multidict==5.1.0
mysql-connector-python==8.0.26
mysqlclient==2.0.3
natsort==7.1.1
numpy==1.21.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.7
pandas-redshift==2.0.5
pandas==1.3.1
paramiko==2.7.2
pendulum==2.1.2
prison==0.1.3
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
psycopg2==2.9.1
pycparser==2.20
pygments==2.9.0
pyinstaller-hooks-contrib==2021.2
pyinstaller==4.5.1
pyjwt==1.7.1
pymssql==2.2.2
pynacl==1.4.0
pyodbc==4.0.31
pyparsing==2.4.7
pyrsistent==0.18.0
pysftp==0.2.9
python-daemon==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pyyaml==5.4.1
ref==0.0.2.2
requests==2.23.0
rfc3986[idna2008]==1.5.0
rich==9.2.0
s3fs==2021.7.0
s3transfer==0.4.2
setproctitle==1.2.2
simple-salesforce==1.11.3
six==1.16.0
sniffio==1.2.0
sqlalchemy-jsonfield==1.0.0
sqlalchemy-utils==0.37.8
sqlalchemy==1.3.24
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thrift==0.13.0
typing-extensions==3.10.0.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
werkzeug==1.0.1
wrapt==1.12.1
wtforms==2.3.3
yarl==1.6.3
zipp==3.5.0
zope.deprecation==4.4.0
zope.interface==5.4.0
